+++ app/src/main/java/com/techcourse/MappingHandler.java
@@ -1,4 +1,4 @@
-package com.interface21.webmvc.servlet.mvc.tobe;
+package com.techcourse;
 
 import jakarta.servlet.http.HttpServletRequest;
